Use search engine name instead of logo
